Misc type conversion warning fixes #rb maxime.mercier #preflight 618d6d32c80d0ce51ae61965
#ROBOMERGE-OWNER: Yoan.StAmant
#ROBOMERGE-AUTHOR: yoan.stamant
#ROBOMERGE-SOURCE: CL 18156659 in //UE5/Main/... via CL 18156710 via CL 18156720
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Staging -> Release-5.0) (v889-18060218)
#ROBOMERGE-CONFLICT from-shelf

[CL 18156988 by Yoan StAmant in 5.0 branch]

+++ Engine/Plugins/Runtime/MassGameplay/Source/MassLOD/Public/MassLODTickRateController.h
@@ -48,7 +48,7 @@ public:
 protected:
 
 	/** Tick rate for each LOD */
-	TStaticArray<int32, EMassLOD::Max> TickRates;
+	TStaticArray<float, EMassLOD::Max> TickRates;
 };
 
 template <typename FVariableTickChunkFragment, typename FLODLogic>
